+++ hatch/commands/install.py
@@ -7,11 +7,10 @@ import click
 from hatch.commands.utils import (
     CONTEXT_SETTINGS, echo_failure, echo_info, echo_success, echo_waiting
 )
-from hatch.config import get_venv_dir
+from hatch.config import get_proper_pip, get_venv_dir
 from hatch.env import install_packages
 from hatch.utils import (
-    NEED_SUBPROCESS_SHELL, ON_WINDOWS, get_admin_command, get_proper_pip,
-    is_project, venv_active
+    NEED_SUBPROCESS_SHELL, ON_WINDOWS, get_admin_command, is_project, venv_active
 )
 from hatch.venv import create_venv, is_venv, venv
