+++ hatch/commands/test.py
@@ -8,10 +8,11 @@ import click
 from hatch.commands.utils import (
     CONTEXT_SETTINGS, echo_failure, echo_info, echo_success, echo_waiting
 )
+from hatch.config import get_proper_python
 from hatch.env import get_editable_package_location, install_packages
 from hatch.utils import (
-    NEED_SUBPROCESS_SHELL, chdir, get_proper_python, get_requirements_file,
-    is_project, resolve_path, venv_active
+    NEED_SUBPROCESS_SHELL, chdir, get_requirements_file, is_project,
+    resolve_path, venv_active
 )
 from hatch.venv import create_venv, is_venv, venv
